Refactoring to enforce consistency for the report

+++ iterations/OptimisedDesign/Config.h
@@ -4,10 +4,10 @@
 struct Config {
 	int img_width;
 	int img_height;
-	float plot_width;
-	float plot_height;
-	float plot_x_min;
-	float plot_y_max;
+	double plot_width;
+	double plot_height;
+	double plot_x_min;
+	double plot_y_max;
 	int max_iteration;
 };
